+++ app/src/main/kotlin/com/revolgenx/anilib/browse/ui/viewmodel/BrowseFilterViewModel.kt
@@ -11,13 +11,13 @@ import com.revolgenx.anilib.common.data.store.GenreCollectionDataStore
 import com.revolgenx.anilib.common.data.store.MediaTagCollectionDataStore
 import com.revolgenx.anilib.common.data.store.ReadableOnCollectionDataStore
 import com.revolgenx.anilib.common.data.store.StreamingOnCollectionDataStore
+import com.revolgenx.anilib.common.ext.get
 import com.revolgenx.anilib.common.ext.launch
 import com.revolgenx.anilib.common.ui.component.menu.MultiSelectModel
 import com.revolgenx.anilib.common.ui.component.menu.SelectFilterModel
 import com.revolgenx.anilib.common.ui.component.menu.SelectType
 import com.revolgenx.anilib.common.ui.viewmodel.BaseViewModel
 import com.revolgenx.anilib.media.ui.model.MediaExternalLinkModel
-import com.revolgenx.anilib.media.ui.model.MediaTagModel
 import kotlinx.coroutines.flow.first
 
 class BrowseFilterViewModel(
@@ -30,65 +30,94 @@ class BrowseFilterViewModel(
 ) : BaseViewModel<BrowseField>() {
 
     override var field: BrowseField by mutableStateOf(BrowseField())
+    val canShowAdultContent = appPreferencesDataStore.displayAdultContent.get()!!
+    val isLoggedIn = appPreferencesDataStore.isLoggedIn.get()
 
-    val canShowAdultContent =
-        mutableStateOf(appPreferencesDataStore.displayAdultContent.get())
+    private var mediaTagCollections =
+        mediaTagCollectionDataStore.data.get().tags.filter { if (canShowAdultContent) true else !it.isAdult }
+    private val genreCollections = genreCollectionDataStore.data.get().genre.filter {
+        if (canShowAdultContent) true else !it.name.contains(
+            "hentai",
+            ignoreCase = true
+        )
+    }
+    private val streamingOnCollections = streamingOnCollectionDataStore.data.get().links
+    private val readableOnCollections = readableOnCollectionDataStore.data.get().links
+
+    private val excludedAnimeTags = mediaTagCollections.filter { it.isExcludedInAnime }
+    private val excludedMangaTags = mediaTagCollections.filter { it.isExcludedInManga }
+    private val excludedAnimeGenre = genreCollections.filter { it.isExcludedInAnime }
+    private val excludedMangaGenre = genreCollections.filter { it.isExcludedInManga }
+
+    val selectMediaTagCollections =
+        mutableStateOf(emptyList<SelectFilterModel<String>>())
     val selectStreamingOnCollections =
         mutableStateOf(emptyList<MultiSelectModel<MediaExternalLinkModel>>())
     val selectReadableOnCollections =
         mutableStateOf(emptyList<MultiSelectModel<MediaExternalLinkModel>>())
-
-    val selectMediaTagCollections =
-        mutableStateOf(emptyList<SelectFilterModel<String>>())
     val selectGenreCollections =
         mutableStateOf(emptyList<SelectFilterModel<String>>())
 
-    private var mediaTagCollections = emptyList<MediaTagModel>()
-    private var genreCollections = emptyList<String>()
-    private var streamingOnCollections = emptyList<MediaExternalLinkModel>()
-    private var readableOnCollections = emptyList<MediaExternalLinkModel>()
+    var isExcludedTagsFiltered = false
+    var isExcludedGenreFiltered = false
 
     init {
-        launch {
-            appPreferencesDataStore.displayAdultContent.data.collect { canShowAdult ->
-                canShowAdultContent.value = canShowAdult
-
-                launch {
-                    mediaTagCollectionDataStore.data.collect { tagCollections ->
-                        mediaTagCollections =
-                            tagCollections.tags.filter { if (canShowAdult!!) true else !it.isAdult }
-                        updateSelectMediaTagCollections()
-                    }
-                }
-
-                launch {
-                    genreCollectionDataStore.data.collect { collections ->
-                        genreCollections =
-                            collections.genre.filter { if (canShowAdult!!) true else it != "Hentai" }
-                        updateSelectGenreCollections()
-                    }
-                }
-            }
-        }
-
-
-        launch {
-            streamingOnCollectionDataStore.data.collect {
-                streamingOnCollections = it.links
-                updateSelectStreamingOnCollections()
-            }
-        }
-        launch {
-            readableOnCollectionDataStore.data.collect {
-                readableOnCollections = it.links
-                updateSelectReadableOnCollections()
-            }
-        }
+        updateSelectMediaTagCollections()
+        updateSelectGenreCollections()
+        updateSelectStreamingOnCollections()
+        updateSelectReadableOnCollections()
     }
 
     fun reloadPrevious() {
         launch {
-            updateField(browseFilterDataStore.data.first().toBrowseField())
+            val previousField = browseFilterDataStore.data.first().toBrowseField()
+            val browseType = previousField.browseType.value
+            if (browseType == BrowseTypes.ANIME || browseType == BrowseTypes.MANGA) {
+                if (!isExcludedTagsFiltered) {
+                    val excludedTags =
+                        if (browseType == BrowseTypes.ANIME) excludedAnimeTags else excludedMangaTags
+                    previousField.tagsIn = previousField.tagsIn?.filter { tagIn ->
+                        excludedTags.find {
+                            it.name.contains(
+                                tagIn,
+                                true
+                            )
+                        } == null
+                    }
+
+                    val tagsNotIn = excludedTags.map { it.name }.toMutableList()
+                    previousField.tagsNotIn?.forEach {
+                        if(!tagsNotIn.contains(it)){
+                            tagsNotIn.add(it)
+                        }
+                    }
+
+                    previousField.tagsNotIn = tagsNotIn
+                }
+
+
+                if (!isExcludedGenreFiltered) {
+                    val excludedGenre =
+                        if (browseType == BrowseTypes.ANIME) excludedAnimeGenre else excludedMangaGenre
+                    previousField.genreIn = previousField.genreIn?.filter { genreIn ->
+                        excludedGenre.find {
+                            it.name.contains(
+                                genreIn,
+                                true
+                            )
+                        } == null
+                    }
+
+                    val genreNotIn = excludedGenre.map { it.name }.toMutableList()
+                    previousField.genreNotIn?.forEach {
+                        if(!genreNotIn.contains(it)){
+                            genreNotIn.add(it)
+                        }
+                    }
+                    previousField.genreNotIn = genreNotIn
+                }
+            }
+            updateField(previousField)
         }
     }
 
@@ -98,8 +127,10 @@ class BrowseFilterViewModel(
 
     fun updateFilter() {
         launch {
-            browseFilterDataStore.updateData {
-                field.toBrowseFilter()
+            if(field.browseType.value == BrowseTypes.ANIME || field.browseType.value == BrowseTypes.MANGA){
+                browseFilterDataStore.updateData {
+                    field.toBrowseFilter()
+                }
             }
         }
     }
@@ -107,6 +138,10 @@ class BrowseFilterViewModel(
     fun updateField(mField: BrowseField) {
         field = mField
         updateUIFilters()
+    }
+
+    fun updateBrowseType(browseTypes: BrowseTypes) {
+        field.browseType.value = browseTypes
     }
 
     private fun updateUIFilters() {
@@ -130,10 +165,10 @@ class BrowseFilterViewModel(
         selectGenreCollections.value =
             genreCollections.map {
                 val included =
-                    if (field.genreIn?.contains(it) == true) SelectType.INCLUDED
-                    else if (field.genreNotIn?.contains(it) == true) SelectType.EXCLUDED
+                    if (field.genreIn?.contains(it.name) == true) SelectType.INCLUDED
+                    else if (field.genreNotIn?.contains(it.name) == true) SelectType.EXCLUDED
                     else SelectType.NONE
-                SelectFilterModel(selected = mutableStateOf(included), data = it)
+                SelectFilterModel(selected = mutableStateOf(included), data = it.name)
             }
     }