+++ app/src/main/kotlin/com/revolgenx/anilib/common/ui/viewmodel/ViewModelModule.kt
@@ -62,15 +62,17 @@ import com.revolgenx.anilib.review.ui.viewmodel.ReviewListFilterViewModel
 import com.revolgenx.anilib.review.ui.viewmodel.ReviewListViewModel
 import com.revolgenx.anilib.review.ui.viewmodel.ReviewViewModel
 import com.revolgenx.anilib.setting.ui.viewmodel.AppearanceSettingsViewModel
+import com.revolgenx.anilib.setting.ui.viewmodel.FilterSettingsViewModel
 import com.revolgenx.anilib.setting.ui.viewmodel.GeneralSettingsViewModel
+import com.revolgenx.anilib.setting.ui.viewmodel.GenreFilterSettingsViewModel
 import com.revolgenx.anilib.setting.ui.viewmodel.MediaListSettingsViewModel
 import com.revolgenx.anilib.setting.ui.viewmodel.MediaSettingsViewModel
 import com.revolgenx.anilib.setting.ui.viewmodel.NotificationSettingsViewModel
 import com.revolgenx.anilib.setting.ui.viewmodel.SettingsViewModel
+import com.revolgenx.anilib.setting.ui.viewmodel.TagFilterSettingsViewModel
 import com.revolgenx.anilib.social.ui.viewmodel.ActivityComposerViewModel
 import com.revolgenx.anilib.social.ui.viewmodel.ActivityInfoViewModel
 import com.revolgenx.anilib.social.ui.viewmodel.ActivityReplyServiceViewModel
-import com.revolgenx.anilib.social.ui.viewmodel.BaseActivityComposerViewModel
 import com.revolgenx.anilib.social.ui.viewmodel.ActivityReplyViewModel
 import com.revolgenx.anilib.social.ui.viewmodel.ActivityUnionFilterViewModel
 import com.revolgenx.anilib.social.ui.viewmodel.ActivityUnionServiceViewModel
@@ -268,7 +270,12 @@ val viewModelModules = module {
     viewModel { parameters -> ActivityInfoViewModel(parameters.get(), get()) }
 
     //browse
-    viewModel { BrowseViewModel(get(), get()) }
+    viewModel { BrowseViewModel(
+        get<Context>().mediaTagCollectionDataStore,
+        get<Context>().genreCollectionDataStore,
+        get(),
+        get()
+    ) }
     viewModel {
         BrowseFilterViewModel(
             get<Context>().mediaTagCollectionDataStore,
@@ -303,7 +310,9 @@ val viewModelModules = module {
     viewModel { MediaSettingsViewModel(get(), get(), get()) }
     viewModel { NotificationSettingsViewModel(get(), get(), get()) }
     viewModel { MediaListSettingsViewModel(get(), get()) }
-
+    viewModel { FilterSettingsViewModel(get()) }
+    viewModel { TagFilterSettingsViewModel(get<Context>().mediaTagCollectionDataStore, get()) }
+    viewModel { GenreFilterSettingsViewModel( get<Context>().genreCollectionDataStore, get()) }
     //explore
     viewModel {
         ExploreMediaViewModel.ExploreTrendingViewModel(