+++ app/src/main/kotlin/com/revolgenx/anilib/media/data/constant/GenreCollection.kt
@@ -1,24 +1,25 @@
 package com.revolgenx.anilib.media.data.constant
 
+import com.revolgenx.anilib.media.ui.model.MediaGenreModel
 
 val genreCollection = listOf(
-    "Action",
-    "Adventure",
-    "Comedy",
-    "Drama",
-    "Ecchi",
-    "Fantasy",
-    "Hentai",
-    "Horror",
-    "Mahou Shoujo",
-    "Mecha",
-    "Music",
-    "Mystery",
-    "Psychological",
-    "Romance",
-    "Sci-Fi",
-    "Slice of Life",
-    "Sports",
-    "Supernatural",
-    "Thriller"
+    MediaGenreModel("Action"),
+    MediaGenreModel("Adventure"),
+    MediaGenreModel("Comedy"),
+    MediaGenreModel("Drama"),
+    MediaGenreModel("Ecchi"),
+    MediaGenreModel("Fantasy"),
+    MediaGenreModel("Hentai"),
+    MediaGenreModel("Horror"),
+    MediaGenreModel("Mahou Shoujo"),
+    MediaGenreModel("Mecha"),
+    MediaGenreModel("Music"),
+    MediaGenreModel("Mystery"),
+    MediaGenreModel("Psychological"),
+    MediaGenreModel("Romance"),
+    MediaGenreModel("Sci-Fi"),
+    MediaGenreModel("Slice of Life"),
+    MediaGenreModel("Sports"),
+    MediaGenreModel("Supernatural"),
+    MediaGenreModel("Thriller")
 )